Raytracing function for scene complete
A lot of warnings to handle in Camera, Scene, and Parser

+++ Raytracer/Scene.h
@@ -93,8 +93,8 @@ public:
 
 	Object ** getAllObjects();
 	Light ** getAllLights();
-	float getNumObjects();
-	float getNumLights();
+	int getNumObjects();
+	int getNumLights();
 };
 
 #endif
